requirements.txt
# Core Telegram Bot API
python-telegram-bot==20.8
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Data Models & Validation
pydantic==2.9.2
pydantic-settings==2.6.1

# Database & Storage
sqlalchemy==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Configuration & Environment
python-dotenv==1.0.1

# HTTP Client & API
httpx==0.27.0
requests==2.32.5
urllib3==2.6.3

# Caching & Rate Limiting
redis==5.0.8
aioredis==2.0.1

# Logging & Monitoring
structlog==24.4.0
prometheus-client==0.21.0

# Security & Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
cryptography==46.0.5

# Async Support
asyncio-mqtt==0.16.2

# JSON Processing
orjson==3.11.7

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pytest-mock==3.14.0

# Development Tools
black==24.10.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0

# Additional Security Updates
starlette==0.52.1
protobuf==7.34.0

# Memory & AI Integration
mem0ai==1.0.3
openai==2.21.0
